ai_api/requirements.txt
google-genai
fastapi
python_dotenv
uvicorn
python-multipart